+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/component/listener/NavigationPanelListener.java
@@ -5,8 +5,8 @@ import javax.inject.Inject;
 import org.vaadin.spring.annotation.VaadinComponent;
 import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.events.EventBus;
-import org.vaadin.spring.events.EventBusScope;
 import org.vaadin.spring.events.EventScope;
+import org.vaadin.spring.events.annotation.EventBusScope;
 import org.vaadin.spring.samples.mvp.ui.component.nav.NavElement;
 import org.vaadin.spring.samples.mvp.ui.component.util.ControlsContext;
