+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/presenter/HeaderPresenter.java
@@ -1,7 +1,7 @@
 package org.vaadin.spring.samples.mvp.ui.presenter;
 
 import org.vaadin.spring.events.Event;
-import org.vaadin.spring.events.EventBusListenerMethod;
+import org.vaadin.spring.events.annotation.EventBusListenerMethod;
 import org.vaadin.spring.navigator.Presenter;
 import org.vaadin.spring.navigator.VaadinPresenter;
 import org.vaadin.spring.samples.mvp.ui.component.util.ControlsContext;
