+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/presenter/MainPresenter.java
@@ -3,8 +3,8 @@ package org.vaadin.spring.samples.mvp.ui.presenter;
 import javax.inject.Inject;
 
 import org.vaadin.spring.events.Event;
-import org.vaadin.spring.events.EventBusListenerMethod;
 import org.vaadin.spring.events.EventScope;
+import org.vaadin.spring.events.annotation.EventBusListenerMethod;
 import org.vaadin.spring.navigator.Presenter;
 import org.vaadin.spring.navigator.VaadinPresenter;
 import org.vaadin.spring.samples.mvp.ui.view.FooterView;
