+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/presenter/NavigationPanelPresenter.java
@@ -3,7 +3,7 @@ package org.vaadin.spring.samples.mvp.ui.presenter;
 import java.util.List;
 
 import org.vaadin.spring.events.Event;
-import org.vaadin.spring.events.EventBusListenerMethod;
+import org.vaadin.spring.events.annotation.EventBusListenerMethod;
 import org.vaadin.spring.navigator.Presenter;
 import org.vaadin.spring.navigator.VaadinPresenter;
 import org.vaadin.spring.samples.mvp.ui.component.nav.NavElement;
