+++ samples/mvp-sample/src/main/java/org/vaadin/spring/samples/mvp/ui/presenter/TabPanelPresenter.java
@@ -1,6 +1,6 @@
 package org.vaadin.spring.samples.mvp.ui.presenter;
 
-import org.vaadin.spring.events.EventBusListenerMethod;
+import org.vaadin.spring.events.annotation.EventBusListenerMethod;
 import org.vaadin.spring.navigator.Presenter;
 import org.vaadin.spring.navigator.VaadinPresenter;
 import org.vaadin.spring.samples.mvp.ui.component.nav.NavElement;
