+++ spring-vaadin-test/src/test/java/org/vaadin/spring/test/ExampleIntegrationTest.java
@@ -26,8 +26,8 @@ import org.vaadin.spring.annotation.EnableVaadin;
 import org.vaadin.spring.annotation.VaadinSessionScope;
 import org.vaadin.spring.annotation.VaadinUIScope;
 import org.vaadin.spring.events.EventBus;
-import org.vaadin.spring.events.EventBusScope;
 import org.vaadin.spring.events.EventScope;
+import org.vaadin.spring.events.annotation.EventBusScope;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
