+++ spring-vaadin/src/main/java/org/vaadin/spring/config/VaadinConfiguration.java
@@ -26,8 +26,8 @@ import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.web.context.request.RequestContextListener;
 import org.vaadin.spring.context.VaadinApplicationContext;
 import org.vaadin.spring.events.EventBus;
-import org.vaadin.spring.events.EventBusScope;
 import org.vaadin.spring.events.EventScope;
+import org.vaadin.spring.events.annotation.EventBusScope;
 import org.vaadin.spring.events.internal.ScopedEventBus;
 import org.vaadin.spring.http.HttpResponseFactory;
 import org.vaadin.spring.http.HttpResponseFilter;
