+++ spring-vaadin/src/main/java/org/vaadin/spring/events/NoEventBusListenerMethodFilter.java
@@ -4,7 +4,7 @@ package org.vaadin.spring.events;
 /**
  * A default filter implementation which always returns true.
  * @author Chris Phillipson (fastnsilver@gmail.com)
- * @see org.vaadin.spring.events.EventBusListenerMethod#scope()
+ * @see org.vaadin.spring.events.annotation.EventBusListenerMethod#scope()
  */
 public class NoEventBusListenerMethodFilter implements EventBusListenerMethodFilter {
